+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 0.6.6 (TBA)
+* New Feature: Added the 'dead man switch' connection feature
 * Improvement: Updated currency list (thanks ilyagalahov / closes #31)
 
 # Version 0.6.5 (27.04.2018)